Ensure that EventStore.ReplayFunc sees a snapshot only
When EventStore.ReplayFunc is called, it should only operate on the
current state of the log, regardless of any transactions that are being
committed after ReplayFunc has been called.

The disk-backed implementation of EventStorage achieves this by
remembering the current file offset when calling ReplayFunc and only
reading events up to that point.