Track when a post was drafted
The view requires this information.

The DraftedAt field is added to PostDraftEvent for this purpose.
The timestamp provided by StoredEvent is not used because the time at
which the event was stored is not the same time at which it occurred.

Furthermore, using the time on which the event was stored would require
all event handlers to work in terms of stored events.  This would mean
crossing the boundary from domain to storage.