Test event store persistence
While writing a thin command line wrapper, I noticed that events do get
appended to the log but on subsequent runs of the program do not appear
when reading from the log.

This turned out to be an error in the filesystem storage backend:  when
reopening the file, the cursor was positioned at the beginning of the
file.

A test is added to check that stored events are visible after reopening
the store.  This required the addition of a `Close` method to EventStore
and EventStorage.  Otherwise there would have been no way in the test to
"reopen" the store.